messaging_app/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-filter==24.2
dj-rest-auth==6.0.0
django-cors-headers==4.3.1
Pillow==10.3.0
python-decouple==3.8
gunicorn==22.0.0
mysqlclient==2.2.4
mysqlclient==2.2.4
